Sema: Fix existential-metatype-to-Any conversion
We were incorrectly unwrapping too many levels of metatype, and
our ErasureExpr would end up with conformances for the instance
type and not the metatype itself.

In the old universe, this was not a problem, but now Any has two
protocol conformance members, so suddently this violated invariants.

This was a regression, introduced in 6027bf46a6915aa976a30cf29ef37c74f74b09fd.

Fixes rdar://125460667.