AST: Fix availability checking for macros as default arguments.
Adoption of the `#isolation` macro as a default argument in the standard
library exposed that the availbility of contents of default argument macro
expansions were being mis-diagnosed. The nodes for auxilary files generated for
default argument expansions were not being added to the enclosing file's
`TypeRefinementContext` hierarchy, so the contents of the expansion would be
diagnosed as if it were located in a context with the minimum availability for
the compilation.

Resolves rdar://125812256