[Demangler] Further optimizations for the remangler.
Replace use of `snprintf()` with some custom code (this is around 10x
faster on my machine).

Move a few of the `Node` member functions to the header so they're
inlined.

Optimize the `deepEquals()` function by adding an `isSimilarTo()`
method on `Node`; the checks that were happening in the `deepEquals()`
function could be implemented more efficiently by making use of
details of the internal representation of `Node`.

rdar://125739630