Pin Ninja checkout to v1.11.1
Ninja 1.12 just released and seems to be dying silently while building
the Swift stdlib. Lets go ahead and pin it to the old tag. Pinned
dependencies are better from a security standpoint anyway.

I'm leaving `next` on the ninja release branch so that we still have
some signal from it, but folks aren't generally living on `next`, so it
should have less impact.