[Distributed] Avoid infinite recursion in distributed thunk on protocol extensions
Because of the new ways we handle distributed requirements to support
distributed resolvable protocols, we now have protocol requirements for
both the thunk and the real method. A distributed thunk inside a
protocol extension now ended up calling "itself" through protocol
witness, but instead should always call the "real" method.

Without this patch such situation would end up in an infinite recurstion
executing such method from executeDistributedTarget when the method was
a default impl provided by extension on a protocol.

This patch resolves this by becoming contextually aware of the
distributed thunk, and allowing it to always call the "real" method.

resolves rdar://125445347