AST: Remove SourceFile::DelayedFunctions
First introduced in 9453e8e4b8d3, then moved to SourceFile in 52c808648f93.

This was a vestigal remnant of the pre-request evaluator declaration checker,
where we had to validate all decls, check conformances, and then check
function bodies because nothing was lazy.

The last remaining purpose was to work around bugs in associated type
inference, which have now been fixed. So we can finally clean up this bit
of technical debt.