Sema: Register the extended type of a package-level extension
The warnings on superfluously public/package imports has a hole where
it could report a package import as not being used by package decls
even if it provides a type extended by a package extension. This is
caused by the exportability checker not being enabled for package
decls, thus not triggering the usual logic registering this reference.

Address this issue by adding a check specifically for package extensions.
We can remove this check once we have exportability checking for package
decls.

rdar://126712864