Sema: Fix regression with -experimental-skip-non-inlinable-function-bodies-without-types
My recent capture analysis refactoring broke a subtle corner case that wasn't
exercised by the test suite.

If a local inside g() was skipped, but the outer function was not skipped, we would
return the empty list of captures for g(). But if the interface type of g() actually
involves an outer generic parameter type, then the empty capture list did not record
the fact that a generic signature was needed, so we attempted to form a call to the
local function without a generic signature.