[6.0][Runtime] Avoid dlsym of objc_isUniquelyReferenced.
We attempted to use the declaration if it exists, and fall back to dlsym. This didn't actually work and we always call dlsym. Include the right header (when available) and add a weak check to the direct call.

rdar://127116080
(cherry picked from commit 4d62c521bdf49ff9566fbe9dd813bb23f5f2c763)