[sil-opt] Add the ability to specify a swift-version.
I implemented this by adding a new option called swift-version. If not set, we
just leave LangOpts.EffectiveLanguageVersion alone and get whatever that default
behavior is. If it is set, we use the same version parsing infrastructure used
by the normal frontend. If the parse fails, we exit -1.

To write the test that I wanted to write to show this behavior, I also needed to
add support to sil-opt for verify-additional-prefixes which is just useful
goodness.