[sil-opt] Hard exit(-1) instead of finishDiagProcessing when failing to setup a compiler invocation.
Running finishDiagProcessing and potentially silently exiting when -verify is
set is the correct behavior for failures later in sil-opt when running Sema or
when running SIL passes. This is incorrect behavior when just setting up the
compiler instance since in such a case, we want to hard exit -1 since this means
that a test has not properly setup the compiler instance itself. It is better to
hard fail and have the test maintainer just update its command line arguments as
appropriate.