[test][Linux] Improve Gold/LLD detection for function_sections.swift
In #72061 the test was modified because LLD and Gold linker map formats
are different, so they need different CHECKs. The original method in
PR #72061 presents problems when the output is not exactly only one
line.

These changes use the first line of the linker map file to decide if the
Gold or the LLD linker map checks should be used, since the format
differs in both for the first line.

This way of checking works in my Linux machine when using the default
linker, when forcing Gold with `-use-ld=gold` and when forcing LLD with
`-use-ld=lld`.